+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
 #### Breaking Changes
 - upgraded to python 3.7 by basing the docker image on divio/base:0.7-py3.7-slim-stretch
-- upgraded to DjangoCMS 3.8 and Django 3.0
+- upgraded to DjangoCMS 3.8 and Django 3.1
 - changed the stage anonymous access url from `https://{base}/?guest-view=true` to `https://{base}/?anonymous-access=true`
 - changed `settings.ALGOLIA_SEARCH_INDEX_TEXT_LIMIT` to `95_000`